Moved the old KMeans in the deprecated module.
