+++ index.js
@@ -26,6 +26,7 @@ if (foundBinary) {
     exports.statistics = require('./src/nodejs/scripts/statistics.js')(pathQmBinary);
     exports.datasets = require('./src/nodejs/datasets/datasets.js')(pathQmBinary);
     exports.qm_util = require('./src/nodejs/scripts/qm_util.js');
+    exports.statistics = require('./src/nodejs/scripts/deprecated.js')(pathQmBinary);
 
 } else {
     throw new Error('qminer binary not found!');
